requirements.txt
Python==3.9.5
Django==3.2.4
djangorestframework==3.12.4
django-admin-honeypot==1.1.0
requests==2.25.1
pytest==6.2.4